Rename preprocessing_multiartifact_python(2)
