+++ src/common/model_utils/tests/test_preprocessing_multiartifact_python.py
@@ -4,7 +4,7 @@ import sys
 
 sys.path.append(str(Path(__file__).parents[2]))  # common/ dir
 
-from model_utils.preprocessing_multiartifact2 import sample_systematic_from_artifacts, sample_windows_from_artifacts, REGEX_PICKLE  # noqa: E402
+from model_utils.preprocessing_multiartifact_python import sample_systematic_from_artifacts, sample_windows_from_artifacts, REGEX_PICKLE  # noqa: E402
 
 
 def test_sample_windows_from_artifacts_multiple_results():
